fix(app): mark and element stacking
This also adjust styling of highlighted marker- and block buttons to be
inline with hovering.

On blur has been replaced with focusout.
